+++ src/main/kotlin/kondorcet/DefaultBallot.kt
@@ -15,3 +15,16 @@ data class DefaultBallot<out T : Any>(override val orderedCandidates: List<Set<T
         fun <T : Any> of(vararg candidates: Collection<T>) = DefaultBallot(candidates.map { it.toSet() })
     }
 }
+
+fun <T : Any> ballot(vararg candidates: T): Ballot<T> = DefaultBallot.of(*candidates)
+fun <T : Any> ballot(vararg candidates: Collection<T>): Ballot<T> = DefaultBallot.of(*candidates)
+fun <T : Any> emptyBallot(): Ballot<T> = DefaultBallot.blank<T>()
+
+operator fun <T : Any> Ballot<T>.plus(ballot: Ballot<T>): Ballot<T> =
+        DefaultBallot(orderedCandidates + ballot.orderedCandidates)
+
+operator fun <T : Any> Ballot<T>.plus(candidates: Collection<T>): Ballot<T> =
+        if (candidates.isEmpty()) this
+        else DefaultBallot(orderedCandidates.plus<Set<T>>(candidates.toSet()))
+
+operator fun <T : Any> Ballot<T>.plus(candidate: T): Ballot<T> = plus(setOf(candidate))