+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 #Change log
 ## 1.0-SNAPSHOT
 ### Added
-* Add `winners` and `winner` properties to `Ballot`
+* `winners` and `winner` properties to `Ballot`
+* Polls
+    * `SimplePoll` simply store ballots
+    * `CheckedPoll` check the validity of the ballots
 
 ### Changed
 * The `Ballot` is now an interface and a `DefaultBallot` class has been added
-* Now the way to create a ballot is `ballot(c1, c2, c3, c4)` or `ballot(setOf(c1, c2), setOf(c3, c4)`
-    (Instead of `Ballot.of(c1, c2, c3)`)
+* Ballots are now created by function call : `ballot(c1, c2, c3, c4)` or `ballot(setOf(c1, c2), setOf(c3, c4)`
+* The `VoteMethod` has been created to separate it from the poll. It can be specified when calling `Poll.result`
 
 ## Fixed
 * The method `Ballot.hasDuplicates` result was the inversed